tg.sh: implement full GIT_ALTERNATE_OBJECT_DIRECTORIES quoting rule
The full rule isn't just whether or not the directory contains a ":"
(or in this case conservatively alternatively a ";") it's actually
also whether or not it starts with a ".  While the computed temporary
directory name really should never start with a quote, never say
never and just include a check for that too since it's really triial
to add another simple check like that.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>