help: improve formatting of plain text and fix typo
The plain text version of the help is available with
`tg help tg`.

To make some of the formatted lists contained in the original
.rst formatted README more readable in plain text mode, some
of the formatting characters are stripped out when the plain
text version is generated.

Add another pattern to catch the ones that were being missed
making sure to maintain the alignment and replace tabs with
spaces for the same reason in the original lists.

Sneak in a typo fix or two.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>