tg-create: align -n with Git and improve --no-edit
Change the "-n" option to mean "--no-commit" to be more
consistent with Git semantics.  Up to now it has also meant
"--no-edit" as well since it was actually a shortcut for
"--no-edit" and "--no-edit" also implied "--no-commit".

Improve the "--no-edit" operation so that it has the same
effect as using EDITOR=: which means it does not suppress
the initial commit by default any more hence the realignment
of "-n" to mean "--no-commit".  (Of course it can be used
together with "-n" or "--no-commit" to continue to produce
that effect if desired.)

Pass the "--no-edit" message through "git stripspace" with
the "-s" option so that the result ends up the same as it
would with the EDITOR=: hack.  No need to drive users to
employing a hack.  Instead just make --no-edit do what it
should have already been doing.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>