testlib: allow die function to accept a message
Generic "FATAL: Unexpected exit with code 1" messages are not
overly helpful for diagnosing problems.

Allow the "die" function to accept an optional message in the
form of one or more arguments and include that in the "FATAL:"
output along with the status code.

This provides the ability to produce more helpful diagnostic
messages although the plain old "die" is still supported with
exactly the same behavior as it previously had.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>