tg: always pass the correct empty blob to awk
If a .topdeps or .topmsg file is missing, the awk code can substitute
a provided blob instead, but it must be told to do so.

Rather than leaving it up to the caller of the top-level run_awk_topgit_msg
and run_awk_topgit_deps routines to always provide the correct empty blob,
always provide it automatically and remove the option at the top-level API
to even specify it.

This eliminates the problem of having some callers remember to do it and
therefore work properly with missing .topdeps/.topmsg files while others
forget and those "tolerated" branches end up being incorrectly ignored.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>