ref_prepare.awk: support arbitrary blob checking
If the 'chkblob' variable is set output a blob verification
and resolution line for the value.  Currently only one value
is supported but that could be enhanced in the future by
splitting on whitespace.

Successful results will turn into four fields with the fourth
being '?' to distinguish it from normal ':' prepare header lines.

Update all the top-level API functions to always pass the empty
blob value to ref_prepare.awk for verification.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>