test-lib-functions-tg: tg_test_create_branches really does do it all
Make tg_test_create_branch(es) able to create a frankenbranch.

A "frankenbranch" is neither bare nor not bare.  In other words, it
has one of either .topdeps or .topmsg but not both.

Very useful for testing corner cases, but painful to create by
hand; hence the newly added support.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>