topgit_recurse.awk: include excess visit counts
Change each "normal" format recurse deps line from this:

  M T L <node> [<parent> <branch> <chain> <names>]

to this:

  M T L V <node> [<parent> <branch> <chain> <names>]

where "V" is the non-negative (often 0) number of excess
visits to the node during this recursion.

It's trivial to provide this information.  It's more
general purpose than just an "is this a repeat?" flag at
no extra cost and it can serve the "is this a repeat?"
function just as well.

It is one extra field to read, but the field does not
require any processing so the cost is practically nonexistent.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>