tg-summary: show annihilated with -v -v -l
Doubling up the verbosity gets annihilated branches to show
up in a --list (aka -l, -t, --terse) output listing.

Since the only way to tell the annihilated branches apart is
via the "(annihilated)" at the end of the synthesized message
it makes sense to only allow this as an extension of formats
that include the branch message in the output.

First up is the --list format and here it is.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>