tg-summary: introduce --with-related option
The --with-related option extends (and implies) --with-deps
but before performing the "--with-deps" operation a preliminary
"--topgit-heads" step is first performed.

When --with-deps is implied and a single branch name has been
specified (either explicitly or implicitly) and it's spelled
EXACTLY either "HEAD" or "@" then the implicit default is promoted
to --with-related instead of --with-deps.  This is a DWIM trick.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>