tg.sh: use correct remote branch list for recurse_deps_internal
Correct stupid typo that was passing the wrong branch list as the list
of existing remote TopGit branches.  The effect was mostly harmless
other than causing unnecessary slowdown.  When a "toggit.remote" was
set it would appear that all TopGit branches had a remote branch and
then they would be checked for "out-of-date"ness but if the remote
branch did not actually exist the check would silently return "not"
as the answer without error but the check in that case should not
have happened at all.

As a result all the updates that should have happened did, but some
unnecessary work was taking place when only some of the local branches
actually had remote "mate"s.

Correct this by passing in the correct file name.  Stupid typo.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>