tg: provide global --pager (-p) option
Not all commands use a pager for output.  It would be incorrect
to suddenly unconditionally switch them over to using a pager
(even the the pager auto-disables unless the output is to a
"terminal").

Instead provide a global --pager (aka -p) option to go with the
global --no-pager option just like Git does.

For example, tg summary can provide a lot of output really fast
now :) if there are a lot of TopGit branches and it's not paged.
Using `tg -p summary` solves the problem now.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>