tg.sh: make needs_update return 1 for out-of-date remote
Fix the regression since 0.19.1 that caused needs_update to return
a 0 status if the only needed update was from a remote branch.

Turns out most clients only care about the output not the actual
status code so the only affected caller was tg export which would
have failed to cancel an export of a branch only out-of-date in
that fashion.

Correct this and make tg export picky again.  The non-picky-about-
out-of-date-remote-heads behavior can always be obtained whenever
desired by simply using the global -u option.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>