tg.sh: enhance needs_update_check to check branch itself
The needs_update function does not actually check the branch
itself to see if it needs an update from its base or remote
head.  This is because it uses recurse_deps which does not
visit the starting node, only its dependencies (recursively).

The needs_update_check function, however, is meant to be a
one-stop-shop for out-of-date checks and therefore needs to
do this check.  Since it already visits the starting node to
check it for "missing" status it can easily do the other checks
too.

However, some clients don't actually want this since they do
their own checks so they can display an out-of-date remote/base
distinctly from out-of-date dependencies (e.g. tg summary).

Therefore allow this checking to be disabled with an option and
allow the flexibility of also preventing dependencies from being
considered "behind" (while still making their parents "behind")
if they are only out-of-date with respect to their bases/remotes.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>