tg-export.sh: construct playground before trying to use it
When collapsing commits, tg export creates a "playground".

Part of this "playground" uses the branch name.

Branch names that are multi-level (e.g. "t/branch") require
construction (creation) of the necessary subdirectories first
before their playground can be used.

This was being done but not always soon enough in some cases.

Correct the problem and simplify the code in the process.  It
could be simplified even further but it's nice to avoid the
unnecessary mkdir calls when not needed.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>