tg-export.sh: contravene collapse collisions
Eliminate any possibility of directory and filename conflicts
while collapsing a commit by using a suffix on the branch name
(^commit) to store the collapsed result.  The parents are already
stored in a "^parents" suffixed file.  This just moves the commit
itself out of a bare file into a suffixed one thereby allowing
a directory of the same name (without the suffix) to exist
simultaneously with the collapsed commit result.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>