tg: avoid bogus missing status for solitary branches
A single solitary TopGit branch that's not referred to by
any other TopGit branches and has no dependencies (either the
.topdeps file is empty, missing or only has annihilated branches
in it), will not show up in the generated dependency edge list
anywhere unless edges to self are requested.

If a branch does not appear in the list of dependency edges anywhere
the recursion algorithm considers it to be "missing" since it appears
to be actually missing.

Avoid this by requesting emission of self edges when generating
the input for the recursion phase.

Add some tests for this problem too.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>