topgit_msg.awk: allow extracting multiple headers
If the extraction expression begins with a "+" then show
all matches with the "pretty" version of the matching
keyword prefixed (using a ": " of course).

Also while in there, make sure that a regex that uses
alternates (i.e. "|") will truly be anchored to the start
and end and treat a line of only one or more spaces/tabs
as ending the headers section (it's explicitly forbidden
to appear in the headers section by the standard).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>