topgit_msg.awk: mostly avoid outputting blank lines
When in only1 or multimatch mode do not output blank lines.

In regular mode a blank line will still be output but only when
the kind, branch name and empty subject explanations have all been
disabled.

These blank lines are maintained in case the output lines need to
match up one-to-one with the input branch lines.  In only1 and
multimatch mode that can't happen normally anyway so that's why
they get the change.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>