tg-export.sh: support --quilt --force and tidy up usage
Make --force work with --quilt.  No surprises, it doesn't remove
anything with --force it just overwrites preexisting files with
the same names as files being output when --force is used (or it
tries to overwrite them anyway).

Move the usage information to the top and split it out into
something that's actually human readable.  Also mention the extra
options that --quilt takes in the usage for easy reference.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>