testlib: support TESTLIB_TEST_TAP_ONLY for sanity
Putting "TESTLIB_TEST_TAP_ONLY = -1" in a config.mak file
(at the top-level) will abolish the insanity output on failures
even when not running in verbose mode.

Also take this opportunity to include the same test number and
description in the failed output when TESTLIB_TEST_TAP_ONLY is
not active.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>