testlib: warn if there's no TG-TEST-SETTINGS file
Only do it for ../bin-wrappers/tg though.

It's not fatal because it could be useful for testing.

When things are behaving unexpectedly it might be because
an upper-level clean and make has been done but only
individual test script files have since been run directly.

Hence the warning message to identify the problem (and
suggest the solution of running `make settings`).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>