tg-revert.sh: minor optimizations
Use --buffer with cat-file when available.

Accelerate showing all unabbreviated refs.

Avoid stepping on ourself while extracting the refs.  In practice
this did not seem to be a problem, but a long pipeline was reading
from the same file it wrote to and replaced which is a no-no and
has now been corrected.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>