tg-export.sh: allow --quilt export to pre-existing empty dir
There's no reason to disallow export to a pre-existing directory
that's empty.  In fact, allow it to contain any number of ".*"
files as well since --quilt never writes files with a leading
"." (Git ref names may not have "." as either the first or last
character of any component of the ref name).

With this change it's enough to remove the previous --quilt
output files in order to re-run the export without needing to
add a `--force` option.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>