tg export: no need for merge or hook setup
While it's true that tg export (in its --collapse and --linearize
modes) is not a read-only operation on the repository, it produces
a non-TopGit branch as its output in those modes.

For this reason there's no point in having it do merge and/or
hook setup since the produced branch (if not using --quilt) isn't
actually a TopGit branch and receives no benefit from such setup.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>