tg-update.sh: make remote updates work again
Since d7125044b657 (tg: disambiguate
refs/heads/refs/remotes/<name>/<branch>, 2017-06-20,Atopgit-0.19.8),
tg update has been failing with an error whenever a local branch
is out of date with respect to it's remote.  Something like this:

    fatal: Not a valid object name refs/heads/:refs/remotes/<remote>/top-bases/<branch>

Correct this regression by recognizing the leading ':' that introduces
full refname dependencies and removing it.  Now anything with a leading
':' gets treated as a full refname and anything without a leading ':'
gets treated as a TopGit branch name -- the original intention of
d7125044b657 in the first place.

Also toggle the four failures this fixes in the test suite from
"test_expect_failure" to "test_expect_success".

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>