tg-update.sh: show correct blocking ref in message
When a new .topdeps dependency is merged in from a remote
and the new dependency does not already exist locally, then
a helpful message provides hints about where to look.

Fix the check so that a blocking local orphan base gets
detected prooperly (add the missing "/").

This problem only affected the helpful warning message
itself, but not the actual update operation.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>