tg.sh: add '-u' option to wayback-updates alias
The "wayback-updates" alias does a fetch --dry-run to show
what updates have occurred since the wayback state.

If a branch has subsequently been checked out and that
branch has been updated then the "git fetch" would be
updating the current HEAD which it does not like at all
even with "--force --dry-run".

Add the "-u" option to disable the check since it is always
a "--dry-run" and the point is to show the updates that
took place without actually changing anything so it's
perfectly safe to use.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>