*.sh: purge use of test '-a' and '-o' ops
Although the test utility's '-a' and '-o' operators are
documented by POSIX, they are marked obsolete and may
not be supported by some implementations.

Furthermore, the semantics (i.e. precedence and order of
evaluation) are not well defined when they are in use.

Therefore expunge use of them in favor of multiple tests
combined with '||' and/or '&&' etc.

Also sneak in a tiny optimization into the tg-export.sh
argument parsing code since it's affected by the change.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>