tg: eliminate use of xargs -0
The xargs "-0" option is NOT specified by POSIX.

Unfortunately, without use of the "-0" option,
handling arbitrary input path names becomes a
bit awkward.

Remove use of the xargs "-0" option in favor of
properly quoting (xargs-style) any troublesome
input filenames.  Fortunately filenames with
embedded '\n' characters do not need to be handled.

At the same time, add the necessary quoting to
one invocation of xargs that was already not using
the "-0" option but would have failed with oddball
input.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>