tg-update.sh: include real branch in fullcmd
The "tg status" command shows the initial full command that was
interrupted when a "tg update" drops out for manual conflict
resolution.

However, if "tg update" was run with no arguments or with the
shortcut "HEAD" or "@" the displayed command can be singularly
unhelpful in determining what the original operation was trying to
update.

Correct this deficiency by munging the "$fullcmd" value to include
the actual TopGit branch name that's being updated so that the "tg
status" command will show it in the case when no arguments were
given or a single "@" or "HEAD" argument was given.

While this does not cover all possible cases, it covers the vast
majority of them and represents a big improvement over just showing
the default "tg update" or "tg update HEAD" as the invoking command
in the "tg status" output.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>