t0001: tolerate failure involving xargs -0
The "xargs -0" command is NOT specified by POSIX.

TopGit no longer makes any use of it and as a result no longer
requires the sanity check that uses it to succeed.

Change the test to "test_tolerate_failure" instead just in case
there's some system out there that doesn't like "xargs -0".

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>