tg-summary.sh: omit superfluous rdeps-once trailing '^'
When showing recursive dependencies using the --rdeps-once mode
(the default --rdeps mode), a dependency that has already had all
its dependents displayed will be shown without dependents but with
a trailing '^' should it be visited again.

However, if it does not actually have any dependents because it's
a "[BASE]" branch or because all its dependencies have since been
annihilated or it's an external non-TopGit branch, there won't be
anything to see underneath it even when it's shown the first time.

In this case the trailing '^' on subsequent showings can be a bit
misleading as it suggests there are dependencies that have been
elided and can be viewed above somewhere, but there aren't because
the branch in question doesn't actually have any.

Detect this situation and omit the trailing '^' to avoid any possible
confusion in these cases.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>