topgit_*_prepare.awk: do not incorrectly spawn misscmd
When both a "missing" value and a "misscmd" have been supplied,
then the "misscmd" will be spawned once just before the first time
the "missing" value is about to be used.

However, if a valid 'blob' check line has been encountered for the
"missing" value the "misscmd" will not be spawned even if the
"missing" value gets used.

At least that's how it was supposed to work.

Correct the semantic errors to make it work properly and avoid the
unnecessary spawning.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>