topgit_deps.awk: strip trailing \r off .topdeps lines
If, somehow, the .topdeps file ends up with \r\n line endings
in it then awk will not strip that \r and the branch name will
be considered invalid because it contains invalid characters.

Tolerate this situation by stripping up to one and only one
trailing \r from the end of the branch name field.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>