t2000: add kludge for ancient broken shells
When using one of the really ancient versions of /bin/sh that some
much older systems have, inside the testing environment only, it
can become confused and think a name is that of a function but then
be unable to find the function and fail with an error.

This disrupts one of the tests.

Since that particular function is not being tested in this case as
all characters involved are always US-ASCII, add a kludgy call to
`sane_unset -f` so that the broken shell stops looking for the
function definition that it somehow lost.

Modern shells will also stop looking for that function definition
(even though they don't lose it) as a result, but since the test
in question isn't testing that functionality, it's irrelevant.

The problem does not occur (even on that broken ancient /bin/sh)
when TopGit is running normally outside the testing environment.

There is no added test for this in t0001 because no simple reproduction
recipe for it has yet become apparent.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>