tg.sh: refresh short ref cache as needed
The ref_exists_rev_short function caches the "--short" version of
a rev_parse operation.

However, the actual abbreviation length requested can change if
core.abbrev changes or an explicit "--short=n" argument gets passed.

Record the extra argument in the cache (first expanding an ambiguous
"--short" to the current value of core.abbrev) and regenerate the
cache entry if the argument differs.

Normally the abbreviation length will not change during a single
run of tg, but temporary directories can be shared in some
circumstances making it theoretically possible.

Storing one extra field that's essentially "free" since it's
passed in and does not need to be computed does not add any extra
overhead.  The value of "core.abbrev" needs to be read once for
the "--short" expansion, but that result is also cached.

This fixes several failing tests which are now toggled to
test_expect_success.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>