tg-revert.sh: honor --hash when listing all refs
When the acceleration was added for `tg revert --list --no-short`
the code inadvertently failed to check for the --hash/--hash-only
option and always output the ref name in that case.

Retain the acceleration while accomodating the --hash/--hash-only
option at the same time.

Update the affected tests to now be test_expect_success and
slightly enhance them to make sure that when --hash/--hash-only
is omitted output consists of more than just the hash.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>