tg-info: support -i and -w options
Add support for -i and -w options to `tg info` with the same semantics
as `tg summary` has.

Although `tg info` never had support for either `-i` or `-w` in the
past, as it has grown a number of additional introspection commands
and a few functions have migrated into it from elsewhere, the lack
of such support can be painfully apparent at times.

Remedy this deficit now.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>