tg: sanitize `git branch --contains` output
Although the `git for-each-ref` command does now have a `--contains`
option that would preform the same function in a more safe manner,
it's not available until Git version 2.7.0 but TopGit only requires
Git version 1.8.5 or later.

We must properly sanitize the output of `git branch --contains`
before using it to make sure there are no accidents.

While unlikely in practice, the various messages it outputs for a
detached HEAD could potentially result in TopGit coming up with the
wrong answer.

Avoid the issue by making sure output is NOT in color and that any
empty output lines or lines containing any space and/or tab characters
are summarily discarded.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>