tg-info.sh: validate non-TopGit committish shortcut answers
Unlike the other `tg info` options, the `--heads` option allows any
committish to be given and defaults to HEAD (even if it's detached
or not a TopGit branch).

This can be helpful in various situations.

However, it can be expensive to compute the answer in the general
case as all the current independent TopGit heads need to be computed
and walked to match against the output of `git branch --contains`.

But, if the committish actually refers to a TopGit branch there's
a nice shortcut of simply navigating the TopGit DAG.

The shortcut doesn't normally work for non-TopGit committishes, but
if the committish has a non-empty `.topdeps` file in its top-level
directory those branches can be used to walk the TopGit DAG instead.

But there's a little gotcha in that case.  That might produce one
or more answers that don't actually contain the given committish.

Therefore, when a non-TopGit committish has been given and the
fast shortcut produces results, validate each result to make sure
it actually contains the given committish before outputting it.

In the event that none of the shortcut results validate fall back
to the slower and more expensive technique that will find a correct
answer if there is one to be found.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>