tg-tag.sh: show correct name in HEAD ref log displays
Although HEAD is often a symbolic ref, it does have its own
ref log.  While `tg tag -g HEAD` was indeed showing the correct
ref log (even for linked working trees), it was displaying the
name as the ref it pointed to in some cases.

Apparently because this command:

    git rev-parse --revs-only --abbrev-ref=strict HEAD --

Returns a different result when HEAD is a symbolic ref depending
on whether or not it's been run from within a linked working tree
or somewhere else.

Yet another linked working tree anomaly.

Compensate for this problem by always displaying "HEAD" when the
actual name of the log file in question is "HEAD".

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>